Feat: Update app according to PR

+++ db/migrate/20230508143317_create_reservations.rb
@@ -1,8 +1,8 @@
 class CreateReservations < ActiveRecord::Migration[7.0]
   def change
     create_table :reservations do |t|
-      t.integer :user_id
-      t.integer :tutor_id
+      t.bigint :user_id
+      t.bigint :tutor_id
       t.date :date
 
       t.timestamps
